requirements.txt
Flask>=2.2.2
pydantic>=1.9.2
google-cloud-storage>=2.5.0
pytest>=7.1.2
pytest-mock>=3.8.2